configure.ac: Allow to install extra files locally.
Introduce a --with-local-install=[no|yes] option, so we can install
translation files and gsettings schemas under the extension directory.
With this, the extension can be installed locally (no root required)
by doing:

./configure --prefix=$HOME/.local --with-local-install=yes

This uses a really ugly hack to change the location of the locale dir;
I haven't touched (yet) anything related to i18n, so any help checking
that it works is really appreciated.